backend/requirements.txt
flask>=2.3.3
flask-socketio>=5.3.6
flask-sqlalchemy>=3.1.1
flask-cors>=4.0.0
python-dotenv>=1.0.0
# Using SQLite instead of PostgreSQL for easier setup
# psycopg2-binary>=2.9.7
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.3.0
python-socketio>=5.10.0
eventlet>=0.33.3
gunicorn>=21.2.0
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0
gevent>=23.9.1
gevent-websocket>=0.10.1 